// src/pages/ArticlesPage/model/types/ArticlePageSchema.ts
import { EntityState } from "@reduxjs/toolkit";
import { Article, ArticleView } from "entities/Article";

export interface ArticlePageSchema extends EntityState<Article> {
  isLoading?: boolean;
  error?: string;
  view: ArticleView;
  // pagination
  page: number;
  limit?: number;
  hasMore: boolean;
}